Crawler

+++ app/Http/Controllers/admin/CategoryControllers.php
@@ -1,55 +1,63 @@
 <?php
 
-namespace App\Http\Controllers;
-use Illuminate\Support\Str;
+namespace App\Http\Controllers\admin;
+
 use App\Components\CategoryRecursive;
+use App\Http\Controllers\Controller;
+use App\Models\admin;
+use App\Models\brand;
 use App\Models\Category;
+use App\Models\product;
+use App\Models\Slider;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
-
-class CategoryController extends Controller
+use Illuminate\Support\Facades\Session;
+use Illuminate\Support\Str;
+class CategoryControllers extends Controller
 {
-    private $category;
-    
-    public function __construct(Category $category){ 
+    private $category,$product,$brand, $slider;
+   
+    public function __construct(Category $category,product $product,brand $brand,Slider $slider){ 
         $this->category = $category;
+        $this->product =  $product;
+        $this->brand =  $brand;
+        $this->slider =  $slider;
     }
-    
     public function index(){  
-        $categories=$this->category::latest()->paginate(5);
+        $categories=$this->category->latest()->paginate(5);
+        $quantityProduct=$this->product->where('product_category_id')->count();
         return view("admin.category.index",compact("categories"));
     }
     // với sự khác biệt là $this->model->where sử dụng một đối tượng model
     //  đã được khởi tạo trước đó, trong khi Model::where sử dụng tên lớp model trực tiếp.
     public function getCategories($parent_id){
+        
         $recursive = new CategoryRecursive($this->category);
         $htmlOption = $recursive->categoryRecursive($parent_id);
         return $htmlOption;
     }
 
     public function create(){ 
+        
           $htmlOption= $this->getCategories("");
           return view("admin.category.add",compact('htmlOption') );
     }
     
     public function store(Request $request, Category $category)
     {
-            $messages = [
-                "category_name.required" => 'Không được để trống',
-                "category_name.unique" => 'Tên danh mục đã tồn tại'
-            ];
-            $request->validate([
-                "category_name" => 'required|unique:categories'
-            ], $messages);
-            
-            $slug = Str::of($request->input('name'))->slug('-');
-            $category = $category->create([
-                'category_name' => $request->input('category_name'),
-                'category_parent_id' => $request->input("category_parent_id")||0,
-                'category_slug' => $slug,
-            ]);
+        $messages = [
+            "city.required" => 'Vui lòng chọn!',
+            "province.required" => 'Vui lòng chọn!',
+            "ward.required" => 'Vui lòng chọn!'
+        ];
+        $rules = [
+            "category_name" => 'required',
+            "province" => 'required',
+            "ward" => 'required'
+        ];
+        $request->validate($rules, $messages); 
             // Gửi thông báo thành công
-            session()->flash('success', 'Thêm danh mục thành công!');
+            // session()->flash('success', 'Thêm danh mục thành công!');
             return back()->with('success', 'Thêm danh mục thành công!');
         
     }
@@ -58,11 +66,14 @@ class CategoryController extends Controller
 
     
     public function edit(Request $request,$id){
+        
         $category=$this->category::find($id);
         $htmlOption= $this->getCategories($category->category_parent_id);
+        
         return view("admin.category.edit",compact("htmlOption","category"));
     }
     public function update(Request $request,$id){
+        
             $request->validate([
                 "category_name"=>'required'
             ],
@@ -80,17 +91,28 @@ class CategoryController extends Controller
             return redirect()->route('category.index'); 
     }
     public function delete($id){
-        try{
-            $this->category->where("category_parent_id",$id)->delete();
-            $this->category->find($id)->delete(); 
-            session()->flash('success', 'Xóa danh mục thành công!');
-            return redirect()->back(); 
+        
+        try{ 
+            $foundChildrenCategory=$this->category->where("category_parent_id",$id)->get();
+           if($foundChildrenCategory->isNotEmpty()){ 
+             return response()->json(['code' => 200, 'message' => 'Vui lòng xóa hết danh mục con!']);
+           }
+           $foundProduct=$this->product->where("product_category_id",$id)->get();
+           if($foundProduct->isNotEmpty()){
+            session()->flash('error', 'Vui lòng xóa hết sản phẩm thuộc danh mục!');
+            return response()->json(['code' => 500, 'message' => 'Vui lòng xóa hết sản phẩm thuộc danh mục này!']);
+           }
+            $this->category->find($id)->delete();  
+            return response()->json(['code' => 200, 'message' => 'Xóa sản phẩm thành công!']); 
             } catch (\Exception $e) {
                 // Log lỗi
                 Log::error($e->getMessage());
                 // Gửi thông báo lỗi
-                session()->flash('error', 'Đã xảy ra lỗi khi xóa danh mục.');
+                return response()->json(['code' => 500, 'message' => 'Đã xảy ra lỗi khi xóa danh mục.']); 
             }
         
     }
+
+
+
 }
